refactor: move extract_zip to utils
