+++ vrc-get-vpm/src/utils/mod.rs
@@ -1,4 +1,5 @@
 mod copy_recursive;
+mod extract_zip;
 
 use async_zip::error::ZipError;
 use futures::stream::FuturesUnordered;
@@ -14,6 +15,7 @@ use std::task::{ready, Context, Poll};
 use tokio::fs::{read_dir, DirEntry, ReadDir};
 
 pub(crate) use copy_recursive::copy_recursive;
+pub(crate) use extract_zip::extract_zip;
 
 pub(crate) trait PathBufExt {
     fn joined(self, into: impl AsRef<Path>) -> Self;
